Create from the Astropy package template
This sets up the packaging from the Astropy package template
(https://github.com/astropy/package-template). A number of
modifications still need to be made because some of the test
infrastructure might not be completely relevant for a Sphinx
infrastructure package, as opposed to an affiliate package that depends
on astropy and sphinx_astropy. On the other hand, at least we're
starting from as canonical a starting point as possible.