Add cron-ci GitHub Actions workflow
The purpose of this workflow is to run weekly (on Monday mornings) and
test the master branch of the codebase against against the master branch
of Sphinx (as configured by the devdeps test factor in Tox). This gives
us a heads-up about possible compatibility issues.

As well, we run the linkcheck test target to give us a heads-up about
any broken links.