Configure sphinx build
The main goals of this configuration is to avoid depending on
sphinx_astropy and its extensions since we can't depend on the upstream
package.

In some cases this has meant porting configuration defaults from
sphinx_astropy.conf.v1.

Since we can't use the edit_on_github extension directly I've removed
the related metadata from setup.cfg. If we re-implement it locally we can
add this back.