Add example directive
Users add the example directive to their documents to mark up content as
an example.

The directive adds the ExampleMarkerNode, which then becomes a <div> in
the HTML build. That div has a unique ID based on the title with an
example-src- prefix. It also has a class of example-index-source.

To test this, we added page-with-examples.rst to the test case and
checked the HTML output from the Sphinx build.

We also test that the directive and node were registered properly with
Sphinx/docutils.