Add preprocess_examples function
The purpose of this function is to run on build-inited, find all
examples in the documentation source, and build out reStructuredText
page stubs for each example (and eventually also the index pages for
those examples and their tags.

This also adds some configuration to let the user control the extension:

- example_index_enabled (turns the index generation on/off)
- example_index_dir (directory where the example index is added)
- example_index_h1 (character to use for title underlines in rst
  for the examplepage template.

We test that sub pages are generated for the known examples in the test
cases.